Move types to shared/types
