+++ backend/questions/question.service.ts
@@ -1,7 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
-import { Question } from '../types/question';
-import { Answer } from '../types/answer';
-import { Comment } from '../types/comment';
+import { Question } from '../../shared/types/question';
+import { Answer } from '../../shared/types/answer';
+import { Comment } from '../../shared/types/comment';
 import writeQuestion from './question.model';
 
 const addNewQuestion = async (
