+++ backend/subjects/subject.service.ts
@@ -1,5 +1,5 @@
-import { Question } from '../types/question';
-import { Subject } from '../types/subject';
+import { Question } from '../../shared/types/question';
+import { Subject } from '../../shared/types/subject';
 import { writeSubject, checkIfSubjectExists } from './subject.model';
 
 const addNewSubject = async (title: string, id: string): Promise<void> => {
